refactor(engine): extract frontmatter post-processing out of QuickAddEngine
The QuickAddEngine base class (322 lines) carried ~230 lines of template
property-type YAML logic — structured-variable validation, front-matter
post-processing, and nested-key assignment — that every engine inherited
whether or not it dealt with structured YAML. That mixed two concerns
(file/folder creation infra vs. YAML property-type handling) into one fat
base.

Extract that concern into a composed collaborator, engine/helpers/
frontmatterPostProcessor.ts, exporting validateStructuredVariables,
shouldPostProcessFrontMatter, postProcessFrontMatter, and assignFrontmatterValue
as free functions taking app. TemplateEngine, CaptureChoiceEngine, and
TemplateInsertEngine now call these directly instead of via inherited methods.
QuickAddEngine drops from 322 to 89 lines, focused on file/folder/path creation.

The validation unit test now imports validateStructuredVariables directly
instead of subclassing the base to reach a protected method.

Behavior-preserving (logic moved verbatim): tsc/lint/build clean, 2127 tests
green — including CaptureChoiceEngine.template-property-types.test.ts which runs
the REAL engine and asserts structured-array YAML output through the extracted
functions. Plugin loads with no errors and a template run verified live.
